requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-filter  # Filtering support
#browsable API
markdown
pyyaml 
uritemplate